requirements.txt
streamlit
numpy
opencv-python
matplotlib
Pillow 